+++ NEWS.md
@@ -2,6 +2,7 @@
 
 ### Breaking changes
 * `panel_group()` and `panel_item()` are deprecated. Please use the `bslib::accordion()` and `bslib::accordion_panel()` instead.
+* `nested_closeable_modal()` was deprecated.
 
 ### Bug fixes
 * Recompute the `live-search` option value dynamically in `updateOptionalSelectInput` (#291)